restructures e2e tests for swap and send redesigns, removes dead code from previous implementation, abstracts common helpers to external files
