+++ extension/e2e-tests/sendPayment.test.ts
@@ -1,13 +1,6 @@
-import { test, expect, expectPageToHaveScreenshot } from "./test-fixtures";
-import {
-  login,
-  loginAndFund,
-  loginToTestAccount,
-  PASSWORD,
-} from "./helpers/login";
+import { test, expect } from "./test-fixtures";
+import { login, loginAndFund, loginToTestAccount } from "./helpers/login";
 import { TEST_M_ADDRESS, TEST_TOKEN_ADDRESS } from "./helpers/test-token";
-import { sendXlmPayment } from "./helpers/sendPayment";
-import { truncatedPublicKey } from "../src/helpers/stellar";
 
 test("Swap doesn't throw error when account is unfunded", async ({
   page,
@@ -18,7 +11,7 @@ test("Swap doesn't throw error when account is unfunded", async ({
 
   await page.getByTestId("nav-link-swap").click();
   await expect(page.getByTestId("AppHeaderPageTitle")).toContainText(
-    "Swap XLM",
+    "Swap from",
   );
 });
 test("Swap shows correct balances for assets", async ({
@@ -167,11 +160,7 @@ test("Swap shows correct balances for assets", async ({
 
   await page.getByTestId("nav-link-swap").click();
   await expect(page.getByTestId("AppHeaderPageTitle")).toContainText(
-    "Swap XLM",
-  );
-  await page.getByText("From").click();
-  await expect(page.getByTestId("AppHeaderPageTitle")).toContainText(
-    "Your assets",
+    "Swap from",
   );
   await expect(page.getByText("100 FOO")).toBeVisible();
   await expect(page.getByText("10 BAZ")).toBeVisible();
@@ -186,15 +175,13 @@ test("Send doesn't throw error when account is unfunded", async ({
   await login({ page, extensionId });
   await page.getByTestId("nav-link-send").click({ force: true });
 
-  await expect(page.getByText("Send To")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
   await page
     .getByTestId("send-to-input")
     .fill("GBTYAFHGNZSTE4VBWZYAGB3SRGJEPTI5I4Y22KZ4JTVAN56LESB6JZOF");
   await page.getByText("Continue").click({ force: true });
 
-  await expect(page.getByTestId("AppHeaderPageTitle")).toContainText(
-    "Send XLM",
-  );
+  await expect(page.getByTestId("AppHeaderPageTitle")).toContainText("Send");
 });
 test("Send doesn't throw error when creating muxed account", async ({
   page,
@@ -204,7 +191,7 @@ test("Send doesn't throw error when creating muxed account", async ({
   await loginAndFund({ page, extensionId });
   await page.getByTestId("nav-link-send").click({ force: true });
 
-  await expect(page.getByText("Send To")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
   await page
     .getByTestId("send-to-input")
     .fill(
@@ -215,14 +202,11 @@ test("Send doesn't throw error when creating muxed account", async ({
   ).toBeVisible();
   await page.getByText("Continue").click();
 
-  await expect(page.getByTestId("AppHeaderPageTitle")).toContainText(
-    "Send XLM",
-  );
+  await expect(page.getByTestId("AppHeaderPageTitle")).toContainText("Send");
+  await page.getByTestId(`SendRow-native`).click({ force: true });
   await page.getByTestId("send-amount-amount-input").fill("1");
-  await page.getByText("Continue").click({ force: true });
-  await expect(page.getByText("Send Settings")).toBeVisible();
-  await page.getByText("Review Send").click();
-  await expect(page.getByText("Confirm Send")).toBeVisible({
+  await page.getByText("Review Send").click({ force: true });
+  await expect(page.getByText("You are sending")).toBeVisible({
     timeout: 200000,
   });
 });
@@ -235,127 +219,41 @@ test("Send XLM payments to recent federated addresses", async ({
   await loginToTestAccount({ page, extensionId });
   await page.getByTestId("nav-link-send").click({ force: true });
 
-  await expect(page.getByText("Send To")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
 
   await page.getByTestId("send-to-input").fill("freighter.pb*lobstr.co");
   await expect(page.getByTestId("send-to-identicon")).toBeVisible();
   await page.getByText("Continue").click({ force: true });
 
-  await expect(page.getByText("Send XLM")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
+  await page.getByTestId(`SendRow-native`).click({ force: true });
   await page.getByTestId("send-amount-amount-input").fill("1");
-  await page.getByText("Continue").click({ force: true });
+  await page.getByText("Review Send").click({ force: true });
 
-  await expect(page.getByText("Send Settings")).toBeVisible();
-  await expect(page.getByTestId("SendSettingsTransactionFee")).toHaveText(
-    /[0-9]/,
-  );
-  // 100 XLM is the default, so likely a sign the fee was not set properly from Horizon
-  await expect(
-    page.getByTestId("SendSettingsTransactionFee"),
-  ).not.toContainText("100 XLM");
-  await expect(page.getByText("Review Send")).toBeEnabled();
-  await page.getByText("Review Send").click();
+  await expect(page.getByText("You are sending")).toBeVisible();
+  await page.getByTestId("SubmitAction").click({ force: true });
 
-  await expect(page.getByText("Confirm Send")).toBeVisible();
-  await expect(page.getByText("XDR")).toBeVisible();
-  await page.getByTestId("transaction-details-btn-send").click();
-
-  await expect(page.getByText("Successfully sent")).toBeVisible({
+  await expect(page.getByText("Sent!")).toBeVisible({
     timeout: 60000,
   });
 
   await page.getByText("Done").click();
-
   await page.getByTestId("nav-link-send").click();
 
-  await expect(page.getByText("Send To")).toBeVisible();
-  await expect(page.getByText("RECENT")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
+  await expect(page.getByText("Recents")).toBeVisible();
   await page.getByTestId("recent-address-button").click();
-  await page.getByText("Continue").click({ force: true });
+  await page.getByTestId(`SendRow-native`).click({ force: true });
 
-  await expect(page.getByText("Send XLM")).toBeVisible();
   await page.getByTestId("send-amount-amount-input").fill("1");
-  await page.getByText("Continue").click({ force: true });
+  await page.getByText("Review Send").click({ force: true });
 
-  await expect(page.getByText("Send Settings")).toBeVisible();
-  await expect(page.getByTestId("SendSettingsTransactionFee")).toHaveText(
-    /[0-9]/,
-  );
-  // 100 XLM is the default, so likely a sign the fee was not set properly from Horizon
-  await expect(
-    page.getByTestId("SendSettingsTransactionFee"),
-  ).not.toContainText("100 XLM");
-  await page.getByText("Review Send").click();
-  await expect(page.getByText("Confirm Send")).toBeVisible();
-  await page.getByTestId("transaction-details-btn-send").click();
-  await expect(page.getByText("Successfully sent")).toBeVisible({
+  await expect(page.getByText("You are sending")).toBeVisible();
+  await page.getByTestId("SubmitAction").click({ force: true });
+
+  await expect(page.getByText("Sent!")).toBeVisible({
     timeout: 60000,
   });
-});
-
-test("Send XLM payments from multiple accounts to G Address", async ({
-  page,
-  extensionId,
-}) => {
-  test.slow();
-  await loginAndFund({ page, extensionId });
-  await sendXlmPayment({ page });
-
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("account-view-account-name").click();
-  await page.getByTestId("add-wallet").click();
-  await page.getByText("Create a new wallet").click();
-
-  // test incorrect password
-  await page.locator("#password-input").fill("wrong password");
-  await page.getByText("Create New Address").click();
-  await expect(page.getByText("Incorrect password")).toBeVisible();
-  await page.locator("#password-input").fill(PASSWORD);
-  await page.getByText("Create New Address").click();
-
-  await expect(page.getByTestId("not-funded")).toBeVisible({
-    timeout: 10000,
-  });
-  await page.getByRole("button", { name: "Fund with Friendbot" }).click();
-
-  await expect(page.getByTestId("account-assets")).toBeVisible({
-    timeout: 30000,
-  });
-  await sendXlmPayment({ page });
-
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("account-view-account-name").click();
-
-  await page.getByText("Account 1").click();
-  await sendXlmPayment({ page });
-
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("BackButton").click();
-  await page.getByTestId("account-view-account-name").click();
-  await page.getByTestId("add-wallet").click();
-  await page.getByText("Import a Stellar secret key").click();
-
-  // test private key account from different mnemonic phrase
-  await page
-    .locator("#privateKey-input")
-    .fill("SDCUXKGHQ4HX5NRX5JN7GMJZUXQBWZXLKF34DLVYZ4KLXXIZTG7Q26JJ");
-  // test incorrect password
-  await page.locator("#password-input").fill("wrongpassword");
-  await page.locator("#authorization-input").click({ force: true });
-
-  await page.getByTestId("import-account-button").click();
-  await expect(
-    page.getByText("Please enter a valid secret key/password combination"),
-  ).toHaveCount(2);
-  await page.locator("#password-input").fill(PASSWORD);
-  await page.getByTestId("import-account-button").click();
-
-  await sendXlmPayment({ page });
 });
 
 test("Send XLM payment to C address", async ({ page, extensionId }) => {
@@ -364,39 +262,24 @@ test("Send XLM payment to C address", async ({ page, extensionId }) => {
 
   // send XLM to C address
   await page.getByTestId("nav-link-send").click({ force: true });
-  await expect(page.getByText("Send To")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
   await page.getByTestId("send-to-input").fill(TEST_TOKEN_ADDRESS);
   await page.getByText("Continue").click({ force: true });
 
-  await expect(page.getByText("Send XLM")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
+  await page.getByTestId(`SendRow-native`).click({ force: true });
   await page.getByTestId("send-amount-amount-input").fill(".001");
-  await page.getByText("Continue").click({ force: true });
+  await page.getByText("Review Send").click({ force: true });
 
-  await expect(page.getByText("Send Settings")).toBeVisible();
-  await expect(page.getByTestId("SendSettingsTransactionFee")).toHaveText(
-    /[0-9]/,
-  );
-  await page.getByPlaceholder("Memo (optional)").fill("Test memo");
-  await page.getByText("Review Send").click();
-
-  await expect(page.getByText("Confirm Send")).toBeVisible({
-    timeout: 200000,
-  });
-  await expectPageToHaveScreenshot({
-    page,
-    screenshot: "send-payment-confirm.png",
-  });
-  await page.getByTestId("transaction-details-btn-send").click();
-
-  await expect(page.getByText("Successfully sent")).toBeVisible({
+  await expect(page.getByText("You are sending")).toBeVisible({
     timeout: 60000,
   });
 
-  await page.getByText("Details").click({ force: true });
+  await page.getByTestId(`SubmitAction`).click({ force: true });
 
-  await expect(page.getByText("Sent XLM")).toBeVisible();
-  await expect(page.getByTestId("asset-amount")).toContainText("0.001");
-  await expect(page.getByTestId("memo")).toContainText("Test memo");
+  await expect(page.getByText("Sent!")).toBeVisible({
+    timeout: 60000,
+  });
 });
 
 test("Send XLM payment to M address", async ({ page, extensionId }) => {
@@ -405,40 +288,24 @@ test("Send XLM payment to M address", async ({ page, extensionId }) => {
 
   // send XLM to C address
   await page.getByTestId("nav-link-send").click({ force: true });
-  await expect(page.getByText("Send To")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
   await page.getByTestId("send-to-input").fill(TEST_M_ADDRESS);
   await page.getByText("Continue").click({ force: true });
 
-  await expect(page.getByText("Send XLM")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
+  await page.getByTestId(`SendRow-native`).click({ force: true });
   await page.getByTestId("send-amount-amount-input").fill(".001");
-  await page.getByText("Continue").click({ force: true });
-
-  await expect(page.getByText("Send Settings")).toBeVisible();
-  await expect(page.getByTestId("SendSettingsTransactionFee")).toHaveText(
-    /[0-9]/,
-  );
   await page.getByText("Review Send").click();
 
-  await expect(page.getByText("Confirm Send")).toBeVisible({
-    timeout: 200000,
-  });
-  await expectPageToHaveScreenshot({
-    page,
-    screenshot: "send-payment-confirm.png",
-  });
-  await page.getByTestId("transaction-details-btn-send").click();
-
-  await expect(page.getByText("Successfully sent")).toBeVisible({
+  await expect(page.getByText("You are sending")).toBeVisible({
     timeout: 60000,
   });
 
-  await page.getByText("Details").click({ force: true });
+  await page.getByTestId(`SubmitAction`).click({ force: true });
 
-  await expect(page.getByText("Sent XLM")).toBeVisible();
-  await expect(page.getByTestId("asset-amount")).toContainText("0.001");
-  await expect(page.getByTestId("to-field")).toContainText(
-    truncatedPublicKey(TEST_M_ADDRESS),
-  );
+  await expect(page.getByText("Sent!")).toBeVisible({
+    timeout: 60000,
+  });
 });
 
 test.skip("Send SAC to C address", async ({ page, extensionId }) => {
@@ -539,7 +406,7 @@ test.skip("Send SAC to C address", async ({ page, extensionId }) => {
   });
 });
 
-test("Send token payment to C address", async ({ page, extensionId }) => {
+test.only("Send token payment to C address", async ({ page, extensionId }) => {
   test.slow();
   await loginToTestAccount({ page, extensionId });
 
@@ -557,28 +424,22 @@ test("Send token payment to C address", async ({ page, extensionId }) => {
   await page.getByTestId("send-to-input").fill(TEST_TOKEN_ADDRESS);
   await page.getByText("Continue").click({ force: true });
 
-  await page.getByTestId("send-amount-asset-select").click({ force: true });
-  await page.getByTestId("Select-assets-row-E2E").click({ force: true });
-
-  await expect(page.getByText("Send E2E")).toBeVisible();
+  await expect(page.getByText("Send")).toBeVisible();
+  await page
+    .getByTestId(`SendRow-E2E:${TEST_TOKEN_ADDRESS}`)
+    .click({ force: true });
   await page.getByTestId("send-amount-amount-input").fill(".001");
-  await page.getByText("Continue").click({ force: true });
+  await page.getByText("Review Send").click({ force: true });
 
-  await expect(page.getByText("Send Settings")).toBeVisible();
-  await expect(page.getByText("Review Send")).toBeEnabled();
-  await page.getByText("Review Send").click();
-
-  await expect(page.getByText("Confirm Send")).toBeVisible();
-  await page.getByTestId("transaction-details-btn-send").click();
-
-  await expect(page.getByText("Successfully sent")).toBeVisible({
-    timeout: 600000,
+  await expect(page.getByText("You are sending")).toBeVisible({
+    timeout: 60000,
   });
 
-  await page.getByText("Details").click({ force: true });
+  await page.getByTestId(`SubmitAction`).click({ force: true });
 
-  await expect(page.getByText("Sent E2E")).toBeVisible();
-  await expect(page.getByTestId("asset-amount")).toContainText("0.001");
+  await expect(page.getByText("Sent!")).toBeVisible({
+    timeout: 60000,
+  });
 });
 test.afterAll(async ({ page, extensionId }) => {
   if (
