--- extension/src/popup/components/sendPayment/SendAmount/AssetSelect/styles.scss
@@ -1,89 +1,0 @@
-.AssetSelect {
-  &__unverified {
-    margin-bottom: 0.75rem;
-  }
-
-  &__wrapper {
-    height: 3.5rem;
-    border: solid 1px var(--sds-clr-gray-06);
-    border-radius: 0.5rem;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    position: relative;
-    cursor: pointer;
-    line-height: 1.75rem;
-    padding: 0.5rem 0.75rem;
-
-    &--path-pay {
-      &:nth-of-type(1) {
-        border-bottom: 0;
-        border-bottom-left-radius: 0;
-        border-bottom-right-radius: 0;
-      }
-      &:nth-of-type(2) {
-        border-top-left-radius: 0;
-        border-top-right-radius: 0;
-      }
-    }
-  }
-
-  &__medium-copy {
-    font-size: 1rem;
-    line-height: 1.5rem;
-    font-weight: var(--font-weight-medium);
-    color: var(--sds-clr-gray-12);
-  }
-
-  &__light-copy {
-    font-size: 0.875rem;
-    line-height: 1rem;
-    font-weight: var(--font-weight-regular);
-    color: var(--sds-clr-gray-12);
-
-    &__label {
-      min-width: 2rem;
-    }
-  }
-
-  &__content {
-    display: flex;
-    flex-direction: row;
-    font-size: 0.875rem;
-    width: 100%;
-
-    &__left {
-      display: flex;
-      align-items: center;
-
-      svg {
-        width: 1rem;
-        height: 1rem;
-        margin: 0 0.5rem;
-      }
-
-      .AccountAssets__asset--logo {
-        margin: 0 0.5rem 0 0.75rem;
-        width: 1.5rem;
-        height: 1.5rem;
-      }
-
-      .AccountAssets__asset--error img {
-        width: 0.75rem;
-        height: 0.75rem;
-      }
-    }
-
-    &__right {
-      flex-grow: 1;
-      display: flex;
-      align-items: center;
-      justify-content: end;
-      svg {
-        height: 1rem;
-        width: 1rem;
-        margin-left: 0.5rem;
-      }
-    }
-  }
-}
