--- extension/src/popup/components/sendPayment/SendAmount/SendType/styles.scss
@@ -1,48 +1,0 @@
-.SendType {
-  &__form-row {
-    display: flex;
-    flex-direction: row;
-    align-items: flex-start;
-  }
-
-  &--label {
-    cursor: pointer;
-    text-transform: none;
-    margin-bottom: 1rem;
-  }
-
-  &__content-wrapper {
-    display: flex;
-    flex-direction: row;
-    align-items: start;
-
-    &__title {
-      display: flex;
-      flex-direction: column;
-    }
-    &__subtext {
-      color: var(--sds-clr-gray-12);
-    }
-  }
-
-  &--radio {
-    &-field {
-      visibility: hidden;
-      width: 0;
-    }
-
-    &-check {
-      display: flex;
-      justify-content: flex-end;
-      flex: 1;
-      visibility: hidden;
-      svg {
-        width: 1rem;
-        height: 1rem;
-      }
-      &--active {
-        visibility: visible;
-      }
-    }
-  }
-}
