+++ extension/src/popup/components/sendPayment/SendAmount/index.tsx
@@ -1,46 +1,24 @@
-import React, { useState, useEffect, useCallback, useRef } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { createPortal } from "react-dom";
 import { Navigate, useLocation } from "react-router-dom";
-import debounce from "lodash/debounce";
 import BigNumber from "bignumber.js";
 import { useFormik } from "formik";
-import {
-  Button,
-  Icon,
-  Loader,
-  Logo,
-  Notification,
-} from "@stellar/design-system";
-import { Asset } from "stellar-sdk";
+import { Button, Icon, Logo } from "@stellar/design-system";
 import { useTranslation } from "react-i18next";
 
-import {
-  AssetSelect,
-  PathPayAssetSelect,
-} from "popup/components/sendPayment/SendAmount/AssetSelect";
 import { LoadingBackground } from "popup/basics/LoadingBackground";
 import { View } from "popup/basics/layout/View";
 import { METRIC_NAMES } from "popup/constants/metricsNames";
 import { AppDispatch } from "popup/App";
 import {
   getAssetFromCanonical,
-  getCanonicalFromAsset,
   isMainnet,
   truncatedPublicKey,
 } from "helpers/stellar";
 import { useNetworkFees } from "popup/helpers/useNetworkFees";
-import { useIsSwap, useIsSoroswapEnabled } from "popup/helpers/useIsSwap";
-import { isAssetSuspicious } from "popup/helpers/blockaid";
 import { emitMetric } from "helpers/metrics";
 import { useRunAfterUpdate } from "popup/helpers/useRunAfterUpdate";
-import {
-  getAssetDecimals,
-  getAvailableBalance,
-  getTokenBalance,
-  isContractId,
-} from "popup/helpers/soroban";
-import { getNativeContractDetails } from "popup/helpers/searchAsset";
+import { getAssetDecimals, getAvailableBalance } from "popup/helpers/soroban";
 import { SubviewHeader } from "popup/components/SubviewHeader";
 import {
   cleanAmount,
@@ -52,34 +30,19 @@ import {
   transactionSubmissionSelector,
   saveAmount,
   saveAsset,
-  saveDestinationAsset,
-  getBestPath,
-  getBestSoroswapPath,
-  getSoroswapTokens,
   saveMemo,
   saveTransactionFee,
   saveTransactionTimeout,
   saveAmountUsd,
 } from "popup/ducks/transactionSubmission";
-import {
-  AccountDoesntExistWarning,
-  shouldAccountDoesntExistWarning,
-} from "popup/components/sendPayment/SendTo";
 import { Loading } from "popup/components/Loading";
-import { ScamAssetWarning } from "popup/components/WarningMessages";
 import { TX_SEND_MAX } from "popup/constants/transaction";
-import { BASE_RESERVE } from "@shared/constants/stellar";
-import { defaultBlockaidScanAssetResult } from "@shared/helpers/stellar";
-import { isSorobanBalance, findAssetBalance } from "popup/helpers/balance";
-import {
-  AssetType,
-  ClassicAsset,
-  LiquidityPoolShareAsset,
-} from "@shared/api/types/account-balance";
+import { findAssetBalance } from "popup/helpers/balance";
 
 import { RequestState, State } from "constants/request";
 import { openTab } from "popup/helpers/navigate";
 import { newTabHref } from "helpers/urls";
+import { AMOUNT_ERROR, InputType } from "helpers/transaction";
 import { reRouteOnboarding } from "popup/helpers/route";
 import { IdenticonImg } from "popup/components/identicons/IdenticonImg";
 import { AssetIcon } from "popup/components/account/AccountAssets";
@@ -93,734 +56,7 @@ import { SimulateTxData } from "./hooks/useSimulateTxData";
 
 import "../styles.scss";
 
-enum AMOUNT_ERROR {
-  TOO_HIGH = "amount too high",
-  DEC_MAX = "too many decimal digits",
-  SEND_MAX = "amount higher than send max",
-}
-
-const ConversionRate = ({
-  source,
-  sourceAmount,
-  dest,
-  destAmount,
-  loading,
-}: {
-  source: string;
-  sourceAmount: string;
-  dest: string;
-  destAmount: string;
-  loading: boolean;
-}) => {
-  const { t } = useTranslation();
-
-  return (
-    <div className="SendAmount__row__rate" data-testid="SendAmountRate">
-      {loading ? (
-        <div data-testid="SendAmountRateLoader">
-          <Loader />
-        </div>
-      ) : (
-        <>
-          {destAmount ? (
-            <span data-testid="SendAmountRateAmount">
-              1 {source} ≈{" "}
-              {new BigNumber(destAmount)
-                .div(new BigNumber(sourceAmount))
-                .toFixed(7)}{" "}
-              {dest}
-            </span>
-          ) : (
-            <span>{t("no path found")}</span>
-          )}
-        </>
-      )}
-    </div>
-  );
-};
-
-// default so can find a path even if user has not given input
-const defaultSourceAmount = "1";
-
-const CHAR_WIDTH = 24;
-
 export const SendAmount = ({
-  goBack,
-  goToNext,
-  goToPaymentType,
-  goToChooseAsset,
-}: {
-  goBack: () => void;
-  goToNext: () => void;
-  goToChooseAsset: () => void;
-  goToPaymentType?: () => void;
-}) => {
-  const { t } = useTranslation();
-  const location = useLocation();
-  const dispatch = useDispatch<AppDispatch>();
-  const runAfterUpdate = useRunAfterUpdate();
-  const spanRef = useRef<HTMLSpanElement>(null);
-  const [inputWidth, setInputWidth] = useState(CHAR_WIDTH);
-
-  const { transactionData, soroswapTokens } = useSelector(
-    transactionSubmissionSelector,
-  );
-  const {
-    amount,
-    asset,
-    destination,
-    destinationAmount,
-    destinationAsset,
-    isToken,
-    destinationIcon,
-    isSoroswap,
-  } = transactionData;
-  const { state: sendAmountData, fetchData } = useGetSendAmountData(
-    {
-      showHidden: false,
-      includeIcons: true,
-    },
-    destination,
-  );
-  const isLoading =
-    sendAmountData.state === RequestState.IDLE ||
-    sendAmountData.state === RequestState.LOADING;
-
-  const isSwap = useIsSwap();
-  const { recommendedFee } = useNetworkFees();
-  const [loadingRate, setLoadingRate] = useState(false);
-  const [showBlockedDomainWarning, setShowBlockedDomainWarning] =
-    useState(false);
-  const [suspiciousAssetData, setSuspiciousAssetData] = useState({
-    domain: "",
-    code: "",
-    issuer: "",
-    image: "",
-    blockaidData: defaultBlockaidScanAssetResult,
-  });
-
-  /* eslint-disable react-hooks/exhaustive-deps */
-  const calculateAvailBalance = useCallback(
-    (selectedAsset: string) => {
-      if (isLoading || sendAmountData.data?.type === AppDataType.REROUTE) {
-        return "";
-      }
-      const userBalances = sendAmountData.data?.userBalances!;
-
-      let _availBalance = new BigNumber("0");
-      const selectedCanonical = getAssetFromCanonical(selectedAsset);
-      const selectedBalance = findAssetBalance(
-        userBalances?.balances,
-        selectedCanonical,
-      );
-      if (isToken && selectedBalance && isSorobanBalance(selectedBalance)) {
-        return getTokenBalance(selectedBalance);
-      }
-      if (userBalances) {
-        // take base reserve into account for XLM payments
-        const minBalance = new BigNumber(
-          (2 + userBalances.subentryCount) * BASE_RESERVE,
-        );
-
-        const balance = selectedBalance?.total || new BigNumber("0");
-        if (selectedAsset === "native") {
-          // needed for different wallet-sdk bignumber.js version
-          const currentBal = new BigNumber(balance.toFixed());
-          _availBalance = currentBal
-            .minus(minBalance)
-            .minus(new BigNumber(Number(recommendedFee)));
-
-          if (_availBalance.lt(minBalance)) {
-            return "0";
-          }
-        } else {
-          // needed for different wallet-sdk bignumber.js version
-          _availBalance = new BigNumber(balance);
-        }
-      }
-
-      return _availBalance.toFixed().toString();
-    },
-    [sendAmountData.data, recommendedFee, isToken],
-  );
-
-  const [availBalance, setAvailBalance] = useState(
-    calculateAvailBalance(asset),
-  );
-
-  const handleContinue = (values: {
-    amount: string;
-    asset: string;
-    destinationAsset: string;
-  }) => {
-    dispatch(saveAmount(cleanAmount(values.amount)));
-    dispatch(saveAsset(values.asset));
-
-    let isDestAssetScam = false;
-
-    if (sendAmountData.data?.type === AppDataType.REROUTE) {
-      return;
-    }
-    const userBalances = sendAmountData.data?.userBalances!;
-    const domains = sendAmountData.data?.domains!;
-    const icons = sendAmountData.data?.icons!;
-
-    const destinationBalance = findAssetBalance(
-      userBalances.balances,
-      getAssetFromCanonical(destinationAsset || "native"),
-    );
-    if (values.destinationAsset) {
-      dispatch(saveDestinationAsset(values.destinationAsset));
-      isDestAssetScam =
-        !!destinationBalance &&
-        "blockaidData" in destinationBalance &&
-        isAssetSuspicious(destinationBalance.blockaidData);
-    }
-    // check for scam asset
-    const assetBalance = findAssetBalance(
-      userBalances.balances,
-      getAssetFromCanonical(asset),
-    );
-    const isSourceAssetScam =
-      !!assetBalance &&
-      "blockaidData" in assetBalance &&
-      isAssetSuspicious(assetBalance.blockaidData);
-    if (isSourceAssetScam) {
-      setShowBlockedDomainWarning(true);
-      setSuspiciousAssetData({
-        code: getAssetFromCanonical(values.asset).code,
-        issuer: getAssetFromCanonical(values.asset).issuer,
-        domain: domains.find(
-          (domain) =>
-            getCanonicalFromAsset(domain.code!, domain.issuer) === values.asset,
-        )!.domain,
-        image: icons[values.asset]!,
-        blockaidData:
-          assetBalance?.blockaidData || defaultBlockaidScanAssetResult,
-      });
-    } else if (isDestAssetScam) {
-      setShowBlockedDomainWarning(true);
-      setSuspiciousAssetData({
-        code: getAssetFromCanonical(values.destinationAsset).code,
-        issuer: getAssetFromCanonical(values.destinationAsset).issuer,
-        domain: domains.find(
-          (domain) =>
-            getCanonicalFromAsset(domain.code!, domain.issuer) ===
-            values.destinationAsset,
-        )!.domain,
-        image: icons[values.destinationAsset]!,
-        blockaidData:
-          !!destinationBalance && "blockaidData" in destinationBalance
-            ? destinationBalance.blockaidData
-            : defaultBlockaidScanAssetResult,
-      });
-    } else {
-      goToNext();
-    }
-  };
-
-  const validate = (values: { amount: string }) => {
-    if (!availBalance) {
-      return {};
-    }
-
-    const val = cleanAmount(values.amount);
-    if (new BigNumber(val).gt(new BigNumber(availBalance))) {
-      return { amount: AMOUNT_ERROR.TOO_HIGH };
-    }
-    if (val.indexOf(".") !== -1 && val.split(".")[1].length > 7) {
-      return { amount: AMOUNT_ERROR.DEC_MAX };
-    }
-    if (new BigNumber(val).gt(new BigNumber(TX_SEND_MAX))) {
-      return { amount: AMOUNT_ERROR.SEND_MAX };
-    }
-    return {};
-  };
-
-  const formik = useFormik({
-    initialValues: { amount, asset, destinationAsset },
-    onSubmit: handleContinue,
-    validate,
-    enableReinitialize: true,
-  });
-
-  useEffect(() => {
-    if (spanRef.current) {
-      // Use scrollWidth to get actual rendered width
-      setInputWidth(spanRef.current.scrollWidth + 2); // Add small padding buffer
-    }
-  }, [formik.values.amount]);
-
-  const showSourceAndDestAsset = !!formik.values.destinationAsset;
-  const parsedSourceAsset = getAssetFromCanonical(formik.values.asset);
-  const parsedDestAsset = getAssetFromCanonical(
-    formik.values.destinationAsset || "native",
-  );
-
-  useEffect(() => {
-    setAvailBalance(calculateAvailBalance(formik.values.asset));
-  }, [calculateAvailBalance, formik.values.asset, sendAmountData.state]);
-
-  // for swaps we're loading and choosing the default destinationAsset here
-  // also, need to check if both source and destination are native
-  useEffect(() => {
-    if (
-      sendAmountData.data?.type === "resolved" &&
-      isSwap &&
-      (!destinationAsset ||
-        (destinationAsset === "native" && asset === "native"))
-    ) {
-      let defaultDestAsset;
-      const userBalances = sendAmountData.data.userBalances;
-
-      // if pre-chosen source asset (eg. from AssetDetails) not XLM, default dest asset to XLM
-      if (formik.values.asset !== Asset.native().toString()) {
-        defaultDestAsset = Asset.native().toString();
-      } else {
-        // otherwise default to first classic side asset if exists
-        const nonXlmAssets = userBalances.balances.filter(
-          (b) =>
-            !("token" in b && b.token.code === "XLM") &&
-            !("liquidityPoolId" in b) &&
-            !("decimals" in b),
-        ) as Exclude<AssetType, LiquidityPoolShareAsset>[];
-        defaultDestAsset =
-          nonXlmAssets && nonXlmAssets[0]
-            ? getCanonicalFromAsset(
-                nonXlmAssets[0].token.code,
-                // casting here since this will be classic or native asset
-                (nonXlmAssets[0] as ClassicAsset).token.issuer?.key,
-              )
-            : Asset.native().toString();
-      }
-
-      dispatch(saveDestinationAsset(defaultDestAsset));
-    }
-  }, [
-    isSwap,
-    dispatch,
-    destinationAsset,
-    sendAmountData.data,
-    formik.values.asset,
-    asset,
-  ]);
-
-  useEffect(() => {
-    if (!soroswapTokens.length) {
-      dispatch(getSoroswapTokens());
-    }
-  }, [isSwap, useIsSoroswapEnabled]);
-
-  useEffect(() => {
-    const getData = async () => {
-      await fetchData();
-    };
-    getData();
-  }, []);
-
-  const getAmountFontSize = () => {
-    const length = formik.values.amount.length;
-    if (length <= 9) {
-      return "";
-    }
-    if (length <= 15) {
-      return "med";
-    }
-    return "small";
-  };
-
-  const db = useCallback(
-    debounce(
-      async (formikAm, sourceAsset, destAsset, publicKey, networkDetails) => {
-        if (isSoroswap) {
-          const getContract = (formAsset: string) =>
-            formAsset === "native"
-              ? getNativeContractDetails(networkDetails).contract
-              : formAsset.split(":")[1];
-
-          await dispatch(
-            getBestSoroswapPath({
-              amount: formikAm,
-              sourceContract: getContract(formik.values.asset),
-              destContract: getContract(formik.values.destinationAsset),
-              networkDetails,
-              publicKey,
-            }),
-          );
-        } else {
-          await dispatch(
-            getBestPath({
-              amount: formikAm,
-              sourceAsset,
-              destAsset,
-              networkDetails,
-            }),
-          );
-        }
-
-        setLoadingRate(false);
-      },
-      2000,
-    ),
-    [],
-  );
-
-  // on asset select get conversion rate
-  useEffect(() => {
-    if (
-      !formik.values.destinationAsset ||
-      Number(formik.values.amount) === 0 ||
-      sendAmountData.state !== RequestState.SUCCESS ||
-      sendAmountData.data.type !== AppDataType.RESOLVED
-    ) {
-      return;
-    }
-    setLoadingRate(true);
-    // clear dest amount before re-calculating for UI
-    db(
-      formik.values.amount || defaultSourceAmount,
-      formik.values.asset,
-      formik.values.destinationAsset,
-      sendAmountData.data.publicKey,
-      sendAmountData.data.networkDetails,
-    );
-  }, [
-    db,
-    formik.values.asset,
-    formik.values.destinationAsset,
-    formik.values.amount,
-    dispatch,
-    sendAmountData.state,
-  ]);
-
-  if (isLoading) {
-    return <Loading />;
-  }
-
-  const hasError = sendAmountData.state === RequestState.ERROR;
-  if (sendAmountData.data?.type === AppDataType.REROUTE) {
-    if (sendAmountData.data.shouldOpenTab) {
-      openTab(newTabHref(sendAmountData.data.routeTarget));
-      window.close();
-    }
-    return (
-      <Navigate
-        to={`${sendAmountData.data.routeTarget}${location.search}`}
-        state={{ from: location }}
-        replace
-      />
-    );
-  }
-
-  if (!hasError) {
-    reRouteOnboarding({
-      type: sendAmountData.data.type,
-      applicationState: sendAmountData.data.applicationState,
-      state: sendAmountData.state,
-    });
-  }
-
-  const sourceBalance = findAssetBalance(
-    sendAmountData.data!.userBalances.balances,
-    parsedSourceAsset,
-  );
-  const destBalance = findAssetBalance(
-    sendAmountData.data!.userBalances.balances,
-    getAssetFromCanonical(formik.values.destinationAsset || "native"),
-  );
-  const sendData = sendAmountData.data!;
-
-  const DecideWarning = () => {
-    // unfunded destination
-    if (
-      !isContractId(destination) &&
-      !isSwap &&
-      shouldAccountDoesntExistWarning(
-        isSwap ? false : sendData.destinationBalances.isFunded!,
-        asset,
-        formik.values.amount || "0",
-      )
-    ) {
-      return <AccountDoesntExistWarning />;
-    }
-    if (formik.errors.amount === AMOUNT_ERROR.TOO_HIGH) {
-      return (
-        <Notification
-          variant="error"
-          title={t("Entered amount is higher than your balance")}
-        />
-      );
-    }
-    if (formik.errors.amount === AMOUNT_ERROR.DEC_MAX) {
-      return (
-        <Notification
-          variant="error"
-          title={`7 ${t("digits after the decimal allowed")}`}
-        />
-      );
-    }
-    if (formik.errors.amount === AMOUNT_ERROR.SEND_MAX) {
-      return (
-        <Notification
-          variant="error"
-          title={`${t(
-            "Entered amount is higher than the maximum send amount",
-          )} (
-          ${formatAmountPreserveCursor(
-            TX_SEND_MAX,
-            formik.values.amount,
-            getAssetDecimals(asset, sendData.userBalances, isToken),
-          )}
-          )`}
-        />
-      );
-    }
-    return null;
-  };
-
-  return (
-    <>
-      {showBlockedDomainWarning &&
-        createPortal(
-          <ScamAssetWarning
-            isSendWarning
-            pillType="Transaction"
-            balances={sendData.userBalances}
-            assetIcons={sendAmountData.data?.icons || {}}
-            domain={suspiciousAssetData.domain}
-            code={suspiciousAssetData.code}
-            issuer={suspiciousAssetData.issuer}
-            image={suspiciousAssetData.image}
-            onClose={() => setShowBlockedDomainWarning(false)}
-            onContinue={goToNext}
-            blockaidData={suspiciousAssetData.blockaidData}
-          />,
-          document.querySelector("#modal-root")!,
-        )}
-      <React.Fragment>
-        <SubviewHeader
-          title={
-            <span>
-              {isSwap ? "Swap" : "Send"}
-              {isSoroswap ? (
-                <span>
-                  on{" "}
-                  <a
-                    href="https://soroswap.finance/"
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    Soroswap
-                  </a>
-                </span>
-              ) : null}
-            </span>
-          }
-          hasBackButton
-          customBackAction={() => {
-            // NOTE: resets base state for transaction data
-            dispatch(saveAsset("native"));
-            dispatch(saveAmount("0"));
-            goBack();
-          }}
-          rightButton={
-            isSwap ? null : (
-              <button
-                onClick={goToPaymentType}
-                className="SendAmount__icon-slider"
-              >
-                <Icon.Expand01 />
-              </button>
-            )
-          }
-        />
-        <View.Content
-          contentFooter={
-            <div className="SendAmount__btn-continue">
-              <Button
-                size="md"
-                disabled={
-                  loadingRate || formik.values.amount === "0" || !formik.isValid
-                }
-                data-testid="send-amount-btn-continue"
-                isFullWidth
-                variant="secondary"
-                onClick={(e) => {
-                  e.preventDefault();
-                  formik.submitForm();
-                }}
-              >
-                {t("Review Send")}
-              </Button>
-            </div>
-          }
-        >
-          <div className="SendAmount">
-            <div className="SendAmount__content">
-              <form>
-                <div className="SendAmount__simplebar__content">
-                  <div className="SendAmount__amount-row">
-                    <div className="SendAmount__amount-input-container">
-                      <span
-                        ref={spanRef}
-                        className={`SendAmount__mirror-amount SendAmount__${getAmountFontSize()}`}
-                      >
-                        {formik.values.amount}
-                      </span>
-                      <input
-                        className={`SendAmount__input-amount ${
-                          isSwap ? "SendAmount__input-amount__full-height" : ""
-                        } SendAmount__${getAmountFontSize()}`}
-                        style={{
-                          width: inputWidth,
-                        }}
-                        data-testid="send-amount-amount-input"
-                        name="amount"
-                        type="text"
-                        placeholder="0"
-                        value={formik.values.amount}
-                        onChange={(e) => {
-                          const input = e.target;
-                          const { amount: newAmount, newCursor } =
-                            formatAmountPreserveCursor(
-                              e.target.value,
-                              formik.values.amount,
-                              getAssetDecimals(
-                                asset,
-                                sendData.userBalances,
-                                isToken,
-                              ),
-                              e.target.selectionStart || 1,
-                            );
-                          formik.setFieldValue("amount", newAmount);
-                          dispatch(saveAmount(newAmount));
-                          runAfterUpdate(() => {
-                            input.selectionStart = newCursor;
-                            input.selectionEnd = newCursor;
-                          });
-                        }}
-                        autoFocus
-                        autoComplete="off"
-                      />
-                      <div
-                        className={`SendAmount__amount-label SendAmount__${getAmountFontSize()}`}
-                      >
-                        {parsedSourceAsset.code}
-                      </div>
-                    </div>
-                  </div>
-                  {showSourceAndDestAsset && formik.values.amount !== "0" && (
-                    <ConversionRate
-                      loading={loadingRate}
-                      source={parsedSourceAsset.code}
-                      sourceAmount={
-                        cleanAmount(formik.values.amount) || defaultSourceAmount
-                      }
-                      dest={parsedDestAsset.code}
-                      destAmount={destinationAmount}
-                    />
-                  )}
-                  <div className="SendAmount__btn-set-max">
-                    <Button
-                      size="md"
-                      variant="tertiary"
-                      isRounded
-                      onClick={() => {
-                        emitMetric(METRIC_NAMES.sendPaymentSetMax);
-                        formik.setFieldValue(
-                          "amount",
-                          calculateAvailBalance(formik.values.asset),
-                        );
-                      }}
-                      data-testid="SendAmountSetMax"
-                    >
-                      {t("Set Max")}
-                    </Button>
-                  </div>
-                  <div
-                    className={`SendAmount__amount-warning${
-                      destinationAsset ? "__path-payment" : ""
-                    }`}
-                  >
-                    <DecideWarning />
-                  </div>
-                  <div className="SendAmount__asset-select-container">
-                    {!showSourceAndDestAsset && (
-                      <AssetSelect
-                        assetCode={parsedSourceAsset.code}
-                        issuerKey={parsedSourceAsset.issuer}
-                        icons={sendAmountData.data?.icons || {}}
-                        isSuspicious={
-                          !!sourceBalance &&
-                          "blockaidData" in sourceBalance &&
-                          isAssetSuspicious(sourceBalance.blockaidData)
-                        }
-                        onSelectAsset={() => {
-                          dispatch(saveAmount("0"));
-                          goToChooseAsset();
-                        }}
-                      />
-                    )}
-                    {showSourceAndDestAsset && (
-                      <>
-                        <PathPayAssetSelect
-                          source={true}
-                          assetCode={parsedSourceAsset.code}
-                          issuerKey={parsedSourceAsset.issuer}
-                          balance={formik.values.amount}
-                          icon=""
-                          icons={sendAmountData.data?.icons || {}}
-                          isSuspicious={
-                            !!sourceBalance &&
-                            "blockaidData" in sourceBalance &&
-                            isAssetSuspicious(sourceBalance.blockaidData)
-                          }
-                          onSelectAsset={() => {
-                            dispatch(saveAmount("0"));
-                            goToChooseAsset();
-                          }}
-                        />
-                        <PathPayAssetSelect
-                          source={false}
-                          assetCode={parsedDestAsset.code}
-                          issuerKey={parsedDestAsset.issuer}
-                          balance={
-                            destinationAmount
-                              ? new BigNumber(destinationAmount).toFixed()
-                              : "0"
-                          }
-                          icon={destinationIcon}
-                          icons={sendAmountData.data!.icons}
-                          isSuspicious={
-                            !!destBalance &&
-                            "blockaidData" in destBalance &&
-                            isAssetSuspicious(destBalance.blockaidData)
-                          }
-                          onSelectAsset={() => {
-                            dispatch(saveAmount("0"));
-                            goToChooseAsset();
-                          }}
-                        />
-                      </>
-                    )}
-                  </div>
-                </div>
-              </form>
-            </div>
-          </div>
-        </View.Content>
-      </React.Fragment>
-      <LoadingBackground
-        onClick={() => {}}
-        isActive={showBlockedDomainWarning}
-      />
-    </>
-  );
-};
-
-type InputType = "crypto" | "fiat";
-
-export const SendAmount2 = ({
   goBack,
   goToNext,
   goToChooseDest,
@@ -850,19 +86,17 @@ export const SendAmount2 = ({
     destination,
   );
 
-  const spanRef = useRef<HTMLSpanElement>(null);
-  const spanRefUsd = useRef<HTMLSpanElement>(null);
   const cryptoInputRef = useRef<HTMLInputElement>(null);
   const usdInputRef = useRef<HTMLInputElement>(null);
 
-  const [inputWidth, setInputWidth] = useState(CHAR_WIDTH);
-  const [inputWidthUsd, setInputWidthUsd] = useState(CHAR_WIDTH);
   const [inputType, setInputType] = useState<InputType>("crypto");
   const [isEditingMemo, setIsEditingMemo] = React.useState(false);
   const [isEditingSettings, setIsEditingSettings] = React.useState(false);
   const [isReviewingTx, setIsReviewingTx] = React.useState(false);
 
   const handleContinue = async () => {
+    const amount = inputType === "crypto" ? formik.values.amount : priceValue!;
+    dispatch(saveAmount(cleanAmount(amount)));
     await fetchSimulationData();
     setIsReviewingTx(true);
   };
@@ -895,18 +129,6 @@ export const SendAmount2 = ({
     sendAmountData.state === RequestState.LOADING;
 
   useEffect(() => {
-    if (spanRef.current) {
-      setInputWidth(spanRef.current.scrollWidth + 8);
-    }
-  }, [formik.values.amount, amount, amountUsd]);
-
-  useEffect(() => {
-    if (spanRefUsd.current) {
-      setInputWidthUsd(spanRefUsd.current.scrollWidth + 8);
-    }
-  }, [formik.values.amountUsd, amount, amountUsd]);
-
-  useEffect(() => {
     if (cryptoInputRef.current) {
       cryptoInputRef.current.focus();
       cryptoInputRef.current.select();
@@ -923,6 +145,7 @@ export const SendAmount2 = ({
       await fetchData();
     };
     getData();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const getAmountFontSize = () => {
@@ -1013,11 +236,6 @@ export const SendAmount2 = ({
     dispatch(saveAmountUsd("0.00"));
     goBack();
   };
-  const onConfirmSend = async () => {
-    const amount = inputType === "crypto" ? formik.values.amount : priceValue!;
-    dispatch(saveAmount(cleanAmount(amount)));
-    goToNext();
-  };
 
   return (
     <React.Fragment>
@@ -1099,17 +317,11 @@ export const SendAmount2 = ({
                   <div className="SendAmount__amount-input-container">
                     {inputType === "crypto" && (
                       <>
-                        <span
-                          ref={spanRef}
-                          className={`SendAmount__mirror-amount SendAmount__${getAmountFontSize()}`}
-                        >
-                          {formik.values.amount}
-                        </span>
                         <input
                           ref={cryptoInputRef}
                           className={`SendAmount__input-amount SendAmount__${getAmountFontSize()}`}
                           style={{
-                            width: inputWidth,
+                            width: `${formik.values.amount.length + 1 || 5}ch`,
                           }}
                           data-testid="send-amount-amount-input"
                           name="amount"
@@ -1148,12 +360,6 @@ export const SendAmount2 = ({
                     )}
                     {inputType === "fiat" && (
                       <>
-                        <span
-                          ref={spanRefUsd}
-                          className={`SendAmount__mirror-amount SendAmount__${getAmountFontSize()}`}
-                        >
-                          {formik.values.amountUsd}
-                        </span>
                         <div
                           className={`SendAmount__amount-label-usd SendAmount__${getAmountFontSize()}`}
                         >
@@ -1163,7 +369,7 @@ export const SendAmount2 = ({
                           ref={usdInputRef}
                           className={`SendAmount__input-amount SendAmount__${getAmountFontSize()}`}
                           style={{
-                            width: inputWidthUsd,
+                            width: `${formik.values.amountUsd.length + 1 || 5}ch`,
                           }}
                           data-testid="send-amount-amount-input"
                           name="amountUsd"
@@ -1360,7 +566,7 @@ export const SendAmount2 = ({
               fee={recommendedFee}
               networkDetails={sendAmountData.data?.networkDetails!}
               onCancel={() => setIsReviewingTx(false)}
-              onConfirm={onConfirmSend}
+              onConfirm={goToNext}
               sendAmount={amount}
               sendPriceUsd={inputType === "crypto" ? priceValue! : amountUsd}
               simulationState={simulationState}
