--- extension/src/popup/components/sendPayment/SendConfirm/TransactionDetails/styles.scss
@@ -1,339 +1,0 @@
-@use "../../../../styles/utils.scss" as *;
-
-.TransactionDetails {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-
-  &__error {
-    flex: 1;
-    padding: var(--View-inset-padding-top);
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-  }
-
-  &__modal-wrapper {
-    display: flex;
-    flex-direction: column;
-  }
-
-  &__send-asset {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    color: var(--sds-clr-gray-12);
-    font-size: 1rem;
-    line-height: 1.5rem;
-
-    &__copy-left {
-      min-width: 0;
-      display: flex;
-      font-weight: var(--font-weight-medium);
-
-      .asset-icon {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-      }
-    }
-
-    &__copy-right {
-      .send-asset-value {
-        display: flex;
-        flex-direction: column;
-      }
-
-      .send-asset-amount {
-        font-size: var(--sds-fs-secondary);
-        color: var(--sds-clr-gray-11);
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        min-width: 0;
-      }
-    }
-  }
-
-  &__loader {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-
-  &__cards {
-    margin-bottom: 1.5rem;
-    row-gap: 1rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    .Card {
-      padding: 1rem;
-    }
-
-    .AccountAssets__asset--logo {
-      height: 1.5rem;
-    }
-
-    .AccountAssets__asset--error img {
-      width: 0.75rem;
-      height: 0.75rem;
-    }
-  }
-
-  &__row {
-    color: rgba(255, 255, 255, 0.6);
-    margin-bottom: 0.75rem;
-    display: flex;
-    justify-content: space-between;
-
-    &__right {
-      color: var(--sds-clr-gray-12);
-      overflow-y: hidden;
-
-      &__item {
-        margin-bottom: 0.25rem;
-      }
-    }
-
-    &__right--hasOverflow {
-      color: var(--sds-clr-gray-12);
-    }
-
-    &__copy {
-      margin-right: pxToRem(4px);
-
-      svg {
-        stroke: var(--sds-clr-gray-09);
-        height: pxToRem(14px);
-        width: pxToRem(14px);
-      }
-    }
-  }
-
-  &__processing {
-    position: absolute;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    background: black;
-    z-index: calc(var(--back--button-z-index) + 1);
-    opacity: 0.8;
-    align-items: center;
-    justify-content: center;
-    top: 0;
-    left: 0;
-
-    &__header {
-      display: flex;
-      color: var(--sds-clr-gray-12);
-      font-weight: var(--font-weight-medium);
-      align-items: center;
-      span {
-        margin-left: 0.5rem;
-      }
-    }
-
-    &__copy {
-      padding-top: 0.5rem;
-      font-size: 0.875rem;
-      line-height: 1.5rem;
-    }
-  }
-
-  &__identicon {
-    display: flex;
-    align-items: center;
-    img {
-      width: 1rem;
-      height: 1rem;
-      border-radius: 1rem;
-      background: var(--color-white);
-      margin-right: 0.5rem;
-    }
-  }
-
-  &__bottom-wrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    position: absolute;
-    top: calc(
-      var(--popup--height) - var(--continue-btn-height) - var(
-          --bottom-btn-margin
-        ) -
-        2.375rem
-    );
-    width: calc(var(--popup--width) - 2 * var(--popup--side-padding));
-
-    > div {
-      display: flex;
-      flex-direction: row;
-    }
-
-    &__buttons {
-      column-gap: 1rem;
-      button {
-        flex: 1;
-      }
-    }
-
-    &__copy {
-      font-size: 0.8125rem;
-      line-height: 1.375rem;
-      min-height: 1.375rem;
-      align-items: center;
-      text-align: center;
-      justify-content: center;
-      padding-top: 1rem;
-    }
-  }
-
-  &__warnings {
-    display: flex;
-    flex-flow: column;
-    margin-top: 1.5rem;
-
-    .WarningMessage__infoBlock {
-      margin-bottom: 0;
-    }
-  }
-}
-
-.TwoAssetCard {
-  display: flex;
-  flex-direction: column;
-  border-radius: 0.5rem;
-  border: 1px solid var(--sds-clr-gray-06);
-  position: relative;
-  margin: 0 0 pxToRem(24px) 0;
-
-  &__row:first-child {
-    border-bottom: 1px solid var(--sds-clr-gray-01);
-  }
-
-  &__row {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 1rem;
-    height: 4.25rem;
-
-    &__left {
-      display: flex;
-      align-items: center;
-      font-weight: var(--font-weight-medium);
-      color: var(--sds-clr-gray-12);
-    }
-
-    &__right {
-      font-weight: var(--font-weight-medium);
-      color: var(--sds-clr-gray-12);
-    }
-  }
-
-  &__arrow-icon {
-    position: absolute;
-    top: calc((100% - 2.5rem) / 2);
-    left: calc((100% - 2.5rem) / 2);
-    width: 2.5rem;
-    height: 2.5rem;
-    border: 1px solid var(--sds-clr-gray-01);
-    border-radius: 2.5rem;
-    z-index: 2;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: var(--sds-clr-gray-01);
-
-    svg {
-      width: 1.5rem;
-      height: 1.5rem;
-    }
-  }
-}
-
-.SendingTransaction {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100%;
-
-  &__Footer {
-    &__Done {
-      margin-top: pxToRem(12px);
-    }
-    &__Subtext {
-      font-size: pxToRem(14px);
-      color: var(--sds-clr-gray-09);
-      margin-bottom: pxToRem(12px);
-    }
-  }
-
-  &__Title {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    margin: pxToRem(32px) 0;
-    font-size: pxToRem(18px);
-
-    .Loader {
-      margin-bottom: pxToRem(8px);
-    }
-
-    &__Success {
-      color: var(--sds-clr-green-09);
-      margin-bottom: pxToRem(8px);
-      width: 2rem;
-      height: 2rem;
-    }
-  }
-
-  &__Summary {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: pxToRem(24px);
-    background-color: var(--sds-clr-gray-03);
-    border-radius: pxToRem(14px);
-
-    &__Description {
-      font-size: pxToRem(16px);
-
-      &__Label {
-        color: var(--sds-clr-gray-09);
-      }
-    }
-
-    &__Assets {
-      display: flex;
-      margin-bottom: pxToRem(28px);
-
-      &__Divider {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin-right: 1rem;
-        svg {
-          color: var(--sds-clr-gray-09);
-        }
-      }
-
-      .IdenticonImg {
-        width: 2rem;
-        height: 2rem;
-        padding: pxToRem(6px);
-        border: 1px solid var(--sds-clr-gray-09);
-        border-radius: pxToRem(100px);
-        margin-right: pxToRem(12px);
-      }
-    }
-  }
-}
