--- extension/src/popup/components/sendPayment/SendSettings/Slippage/styles.scss
@@ -1,35 +1,0 @@
-.Slippage {
-  &__cards {
-    margin-bottom: 0.75rem;
-    display: flex;
-    flex-direction: row;
-    flex-shrink: 1;
-    justify-content: space-between;
-    gap: 0.75rem;
-
-    .Card {
-      cursor: pointer;
-      font-weight: var(--font-weight-medium);
-      font-size: 1rem;
-      height: 2.9rem;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-  }
-
-  &--radio-label {
-    width: 100%;
-  }
-
-  &--radio-field {
-    display: none;
-    &:checked + .Card {
-      border: solid 1px;
-    }
-  }
-
-  &__custom-input {
-    margin-bottom: 2rem;
-  }
-}
