--- extension/src/popup/components/sendPayment/SendSettings/TransactionFee/styles.scss
@@ -1,8 +1,0 @@
-.TransactionFee {
-  &__row {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: 1.5rem;
-  }
-}
