+++ extension/src/popup/components/swap/SwapAsset/styles.scss
