+++ extension/src/popup/ducks/transactionSubmission.ts
@@ -129,6 +129,7 @@ export const submitFreighterTransaction = createAsyncThunk<
         };
         const res = await fetch(`${INDEXER_URL}/submit-tx`, options);
         const response = await res.json();
+        console.log(response);
 
         if (!res.ok) {
           return thunkApi.rejectWithValue({
@@ -138,6 +139,7 @@ export const submitFreighterTransaction = createAsyncThunk<
         }
         return response;
       } catch (e) {
+        console.log(e);
         const message = e instanceof Error ? e.message : JSON.stringify(e);
         return thunkApi.rejectWithValue({
           errorMessage: message,
